# openapi_resolver_ia_serpapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resolver IA API",
        "version": "2.1.0",
        "description": "Endpoints de Resolver IA con integraci\u00f3n a SerpApi: Google, Noticias, Scholar"
    },
    "paths": {
        "/buscar/google": {
            "get": {
                "summary": "Buscar en Google",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "T\u00e9rmino de b\u00fasqueda"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de b\u00fasqueda de Google"
                    }
                }
            }
        },
        "/buscar/noticias": {
            "get": {
                "summary": "Buscar noticias en Google News",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "T\u00e9rmino de b\u00fasqueda de noticias"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de noticias"
                    }
                }
            }
        },
        "/buscar/scholar": {
            "get": {
                "summary": "Buscar en Google Scholar",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "T\u00e9rmino de b\u00fasqueda acad\u00e9mica"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados acad\u00e9micos de Google Scholar"
                    }
                }
            }
        },
        "/analyze": {
            "post": {
                "summary": "Analizar contenido de archivo",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Resultado del an\u00e1lisis de archivo"
                    }
                }
            }
        }
    }
}